Split players/targets into files
**This Commit**
Started as an attempt to pull out some plugins but I didn't know how to
do that with system sets. Instead I just settled for extracting some
stuff into some modules.

**Why?**
Maybe it's easier to read. I'm not really sure. Probably a mistake I'll
regret later. I know the keys stuff will have to come out at some point
since we'll want legends on the UI when I figure that out. The utils
module is also a mess, of course, but that'll probably need to wait
until I understand how to do levels.
